requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
Django==2.1.8
django-cors-headers==3.0.1
django-filter==2.1.0
django-localflavor==2.1
django-webpack-loader==0.6.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
httpie==1.0.2
idna==2.8
mysqlclient==1.4.2.post1
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
